+++ src/dialogs/save_file_dialog.rs
@@ -150,7 +150,7 @@ where
         self.cancel.render(graphics, mouse_xy);
     }
 
-    fn on_key_up(&mut self, key: VirtualKeyCode, _: &Vec<&VirtualKeyCode>) {
+    fn on_key_up(&mut self, key: VirtualKeyCode, _: Coord, _: &Vec<&VirtualKeyCode>) {
         self.name_field.on_key_press(key);
         self.current_dir_field.on_key_press(key);
     }